backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "yevgeny",email = "yyurgenson@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "openai (>=1.77.0,<2.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "psycopg (>=3.2.7,<4.0.0)",
    "fhir-resources (>=8.0.0,<9.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false